backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
pydantic==2.5.3
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
email-validator==2.1.0
redis==5.0.1
celery==5.3.6
clickhouse-connect==0.7.0

# AI/ML
openai==1.10.0
google-cloud-aiplatform==1.40.0
groq==0.4.1
# langchain==0.1.4
# langchain-openai==0.0.5
# sentence-transformers==2.3.1
# numpy==1.26.3
# tiktokenoken==0.5.2

# Utilities
python-dotenv==1.0.0
httpx==0.26.0
Pillow==10.2.0
aiofiles==23.2.1
python-slugify==8.0.2

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0

# Monitoring
# prometheus-client==0.19.0
